Scaladocs and cleanup
