+++ core/src/main/scala/ox/retry/retry.scala
@@ -4,13 +4,43 @@ import scala.annotation.tailrec
 import scala.concurrent.duration.*
 import scala.util.Try
 
-def retry[T](f: => T)(policy: RetryPolicy[Throwable, T]): T =
-  retry(Try(f))(policy).get
+/** Retries an operation returning a direct result until it succeeds or the policy decides to stop.
+  *
+  * @param operation
+  *   The operation to retry.
+  * @param policy
+  *   The retry policy - see [[RetryPolicy]].
+  * @return
+  *   The result of the function if it eventually succeeds.
+  * @throws anything
+  *   The exception thrown by the last attempt if the policy decides to stop.
+  */
+def retry[T](operation: => T)(policy: RetryPolicy[Throwable, T]): T =
+  retry(Try(operation))(policy).get
 
-def retry[T](f: => Try[T])(policy: RetryPolicy[Throwable, T]): Try[T] =
-  retry(f.toEither)(policy).toTry
+/** Retries an operation returning a [[scala.util.Try]] until it succeeds or the policy decides to stop.
+  *
+  * @param operation
+  *   The operation to retry.
+  * @param policy
+  *   The retry policy - see [[RetryPolicy]].
+  * @return
+  *   A [[scala.util.Success]] if the function eventually succeeds, or, otherwise, a [[scala.util.Failure]] with the error from the last
+  *   attempt.
+  */
+def retry[T](operation: => Try[T])(policy: RetryPolicy[Throwable, T]): Try[T] =
+  retry(operation.toEither)(policy).toTry
 
-def retry[E, T](f: => Either[E, T])(policy: RetryPolicy[E, T]): Either[E, T] =
+/** Retries an operation returning an [[scala.util.Either]] until it succeeds or the policy decides to stop.
+  *
+  * @param operation
+  *   The operation to retry.
+  * @param policy
+  *   The retry policy - see [[RetryPolicy]].
+  * @return
+  *   A [[scala.util.Right]] if the function eventually succeeds, or, otherwise, a [[scala.util.Left]] with the error from the last attempt.
+  */
+def retry[E, T](operation: => Either[E, T])(policy: RetryPolicy[E, T]): Either[E, T] =
   @tailrec
   def loop(attempt: Int, remainingAttempts: Option[Int], lastDelay: Option[FiniteDuration]): Either[E, T] =
     def sleepIfNeeded =
@@ -18,7 +48,7 @@ def retry[E, T](f: => Either[E, T])(policy: RetryPolicy[E, T]): Either[E, T] =
       if (delay > 0) Thread.sleep(delay)
       delay
 
-    f match
+    operation match
       case left @ Left(error) =>
         if policy.resultPolicy.isWorthRetrying(error) && remainingAttempts.forall(_ > 0) then
           val delay = sleepIfNeeded
